requirements.txt
ddddocr == 1.1.0
pyyaml == 6.0
requests
